requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.21.0
torch>=1.8.0
torchvision>=0.9.0
pillow>=9.0.0
matplotlib>=3.5.0
easyocr>=1.7.0
pymongo>=4.0.0
flask>=2.0.0
twilio>=7.0.0
python-dotenv>=0.19.0
requests>=2.25.0
werkzeug>=2.0.0
